# docs/api-docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "中间层模块",
        "description": "虚拟宠物医院API文档，前端用。字段默认值：int为-1，String为'This field is currently Empty.'，Date为null，应该会表现为空字符串。",
        "version": "1.0.1-SNAPSHOT"
    },
    "servers": [
        {
            "url": "http://pikapikpinoimac:8090",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/system/archive": {
            "get": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "获取档案信息列表接口",
                "operationId": "getArchives",
                "parameters": [
                    {
                        "name": "storeTime",
                        "in": "query",
                        "description": "记录时间",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "diseaseType",
                        "in": "query",
                        "description": "疾病类型",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "petType",
                        "in": "query",
                        "description": "宠物类型",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "petName",
                        "in": "query",
                        "description": "宠物名字",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "petSex",
                        "in": "query",
                        "description": "宠物性别",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ownerTel",
                        "in": "query",
                        "description": "主人联系电话",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "archiveList": [
                                            {
                                                "petName": "Tom",
                                                "petSex": "M",
                                                "petType": "cat",
                                                "storeTime": "2022-12-31 11:11:11",
                                                "ownerTel": "12345678910",
                                                "diseaseType": "Psychosis",
                                                "archiveId": 1
                                            }
                                        ]
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "更改档案信息接口",
                "operationId": "updateArchive",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "json",
                                "description": "更改后的档案信息",
                                "example": {
                                    "archiveId": 2,
                                    "storeTime": "2022-12-31 11:11:11",
                                    "diseaseType": "Psychosis",
                                    "petType": "cat",
                                    "petName": "Tom",
                                    "petSex": "M",
                                    "ownerTel": "12345678910"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "petName": "Tom",
                                        "petSex": "M",
                                        "petType": "cat",
                                        "storeTime": "2022-12-31 12:22:09",
                                        "ownerTel": "12345678910",
                                        "diseaseType": "Psychosis",
                                        "archiveId": 1
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "新增档案信息接口",
                "operationId": "addArchive",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "json",
                                "description": "新增的档案信息",
                                "example": {
                                    "storeTime": "2022-12-31 12:12:12",
                                    "diseaseType": "Psychosis",
                                    "petType": "cat",
                                    "petName": "Tom",
                                    "petSex": "M",
                                    "ownerTel": "12345678910"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "petName": "Tom",
                                        "petSex": "M",
                                        "petType": "cat",
                                        "storeTime": "2022-12-31 12:12:12",
                                        "ownerTel": "12345678910",
                                        "diseaseType": "Psychosis",
                                        "archiveId": 2
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/system/archive/{archiveId}": {
            "delete": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "删除档案信息接口",
                "operationId": "deleteArchive",
                "parameters": [
                    {
                        "name": "archiveId",
                        "in": "path",
                        "description": "档案信息Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "petName": "Tom",
                                        "petSex": "M",
                                        "petType": "cat",
                                        "storeTime": "2022-12-31 12:12:12",
                                        "ownerTel": "12345678910",
                                        "diseaseType": "Psychosis",
                                        "archiveId": 2
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}